backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.26.4
tensorflow-cpu==2.16.1
scikit-learn==1.6.1
gunicorn==22.0.0